requirements.txt
fastapi
uvicorn
typer
sqlmodel
gitpython
langchain
langchain-ollama
watchdog
pydantic
